+++ src/schemathesis/runner/impl/threadpool.py
@@ -19,7 +19,7 @@ from ...models import CheckFunction, TestResultSet
 from ...stateful import Feedback, Stateful
 from ...targets import Target
 from ...transports.auth import get_requests_auth
-from ...types import RawAuth, RequestCert
+from ...types import RawAuth
 from ...utils import capture_hypothesis_output
 from .. import events
 from .core import BaseRunner, asgi_test, get_session, handle_schema_error, network_test, run_test, wsgi_test
@@ -227,9 +227,6 @@ class ThreadPoolRunner(BaseRunner):
     """Spread different tests among multiple worker threads."""
 
     workers_num: int = 2
-    request_tls_verify: bool | str = True
-    request_proxy: str | None = None
-    request_cert: RequestCert | None = None
 
     def _execute(
         self, results: TestResultSet, stop_event: threading.Event
@@ -333,10 +330,7 @@ class ThreadPoolRunner(BaseRunner):
             "stateful_recursion_limit": self.stateful_recursion_limit,
             "data_generation_methods": self.schema.data_generation_methods,
             "kwargs": {
-                "request_timeout": self.request_timeout,
-                "request_tls_verify": self.request_tls_verify,
-                "request_proxy": self.request_proxy,
-                "request_cert": self.request_cert,
+                "request_config": self.request_config,
                 "store_interactions": self.store_interactions,
                 "max_response_time": self.max_response_time,
                 "dry_run": self.dry_run,